Add hashmap class with some basic methods
The hashmap includes the set(), which adds key-value pair to the hash table,
get() which take the given key and return it, clear() clears the hash table,
length() return the size of the hash map, hash() which hashes the given key.
Also, implemented the keys() which return an array of keys, and values() which
return an array of values. Lastly, the class has has() and entries() method,
which check if the given key is in the hash map and returns all the key-value
pairs, respectively.